+++ Tests/Wire/Resources/Tracking/Sessions/SessionsClientWireTests.swift
@@ -32,6 +32,21 @@ import Chrt
                       "created_at_timestamp": "2024-01-15T09:30:00Z",
                       "terminated": true,
                       "terminated_at_timestamp": "2024-01-15T09:30:00Z",
+                      "last_seen_at_location": {
+                        "type": "Feature",
+                        "geometry": {
+                          "geometries": [
+                            {
+                              "coordinates": [
+                                []
+                              ],
+                              "type": "LineString"
+                            }
+                          ],
+                          "type": "GeometryCollection"
+                        }
+                      },
+                      "last_seen_at_timestamp": "2024-01-15T09:30:00Z",
                       "fa_alert_ids": [
                         1
                       ],
@@ -81,6 +96,25 @@ import Chrt
                     createdAtTimestamp: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
                     terminated: Optional(true),
                     terminatedAtTimestamp: Optional(try! Date("2024-01-15T09:30:00Z", strategy: .iso8601)),
+                    lastSeenAtLocation: Optional(LocationFeature(
+                        type: .feature,
+                        geometry: .geometryCollection(
+                            .init(
+                                geometries: [
+                                    .lineString(
+                                        .init(
+                                            coordinates: [
+                                                LineStringCoordinatesItem.position2D(
+                                                    []
+                                                )
+                                            ]
+                                        )
+                                    )
+                                ]
+                            )
+                        )
+                    )),
+                    lastSeenAtTimestamp: Optional(try! Date("2024-01-15T09:30:00Z", strategy: .iso8601)),
                     faAlertIds: Optional([
                         1
                     ]),
@@ -112,8 +146,11 @@ import Chrt
             filterFlightNumber: "filter_flight_number",
             filterFaFlightId: "filter_fa_flight_id",
             filterFlightLoadedStatus: "filter_flight_loaded_status",
+            filterHasLastSeen: true,
             filterCreatedAtTimestampGte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
             filterCreatedAtTimestampLte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
+            filterLastSeenAtTimestampGte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
+            filterLastSeenAtTimestampLte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
             filterTerminationScheduledForTimestampGte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
             filterTerminationScheduledForTimestampLte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
             filterTerminatedAtTimestampGte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
@@ -150,6 +187,31 @@ import Chrt
                   "created_at_timestamp": "2024-01-15T09:30:00Z",
                   "terminated": true,
                   "terminated_at_timestamp": "2024-01-15T09:30:00Z",
+                  "last_seen_at_location": {
+                    "bbox": [
+                      {
+                        "key": "value"
+                      }
+                    ],
+                    "type": "Feature",
+                    "geometry": {
+                      "geometries": [
+                        {
+                          "coordinates": [
+                            []
+                          ],
+                          "type": "LineString"
+                        }
+                      ],
+                      "type": "GeometryCollection"
+                    },
+                    "properties": {
+                      "address": "address",
+                      "name": "name"
+                    },
+                    "id": 1
+                  },
+                  "last_seen_at_timestamp": "2024-01-15T09:30:00Z",
                   "fa_alert_ids": [
                     1
                   ],
@@ -197,6 +259,39 @@ import Chrt
             createdAtTimestamp: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
             terminated: Optional(true),
             terminatedAtTimestamp: Optional(try! Date("2024-01-15T09:30:00Z", strategy: .iso8601)),
+            lastSeenAtLocation: Optional(LocationFeature(
+                bbox: Optional([
+                    JSONValue.object(
+                        [
+                            "key": JSONValue.string("value")
+                        ]
+                    )
+                ]),
+                type: .feature,
+                geometry: .geometryCollection(
+                    .init(
+                        geometries: [
+                            .lineString(
+                                .init(
+                                    coordinates: [
+                                        LineStringCoordinatesItem.position2D(
+                                            []
+                                        )
+                                    ]
+                                )
+                            )
+                        ]
+                    )
+                ),
+                properties: Optional(LocationProperties(
+                    address: Optional("address"),
+                    name: Optional("name")
+                )),
+                id: Optional(Id.int(
+                    1
+                ))
+            )),
+            lastSeenAtTimestamp: Optional(try! Date("2024-01-15T09:30:00Z", strategy: .iso8601)),
             faAlertIds: Optional([
                 1
             ]),